+++ src/ocaml_version.ml
@@ -1,5 +1,7 @@
 type t = int * int * int
 
+let make x = x
+
 let of_ocaml_config ocfg =
   Ocaml_config.version ocfg
 
@@ -23,3 +25,9 @@ let supports_response_file version =
 
 let ocamlmklib_supports_response_file version =
   version >= (4, 08, 0)
+
+let pervasives_includes_result version =
+  version >= (4, 03, 0)
+
+let stdlib_includes_uchar version =
+  version >= (4, 03, 0)